+++ MobileGarage/wearApp/src/debug/java/com/chriscartland/garage/wear/debug/ScreenshotStagesActivity.kt
@@ -64,6 +64,7 @@ class ScreenshotStagesActivity : ComponentActivity() {
                         ),
                         buttonState = fixture.buttonState,
                         isHolding = false,
+                        signInError = false,
                         onDoorTap = {},
                         onHoldStart = {},
                         onHoldEnd = {},